+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/JeiHelper.java
@@ -2,8 +2,11 @@ package com.refinedmods.refinedstorage.jei.common;
 
 import mezz.jei.api.runtime.IJeiRuntime;
 
-public class JeiProxy {
-    public String getSearchFieldText() {
+public final class JeiHelper {
+    private JeiHelper() {
+    }
+
+    public static String getSearchFieldText() {
         final IJeiRuntime runtime = RefinedStorageModPlugin.getRuntime();
         if (runtime == null) {
             return "";
@@ -11,7 +14,7 @@ public class JeiProxy {
         return runtime.getIngredientFilter().getFilterText();
     }
 
-    public void setSearchFieldText(final String text) {
+    public static void setSearchFieldText(final String text) {
         final IJeiRuntime runtime = RefinedStorageModPlugin.getRuntime();
         if (runtime != null) {
             runtime.getIngredientFilter().setFilterText(text);