+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/RefinedStorageModPlugin.java
@@ -2,25 +2,19 @@ package com.refinedmods.refinedstorage.jei.common;
 
 import javax.annotation.Nullable;
 
-import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
-import com.refinedmods.refinedstorage2.platform.api.support.resource.RecipeModIngredientConverter;
 import com.refinedmods.refinedstorage2.platform.common.grid.screen.AbstractGridScreen;
 import com.refinedmods.refinedstorage2.platform.common.support.AbstractBaseScreen;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
 import mezz.jei.api.constants.RecipeTypes;
-import mezz.jei.api.helpers.IPlatformFluidHelper;
 import mezz.jei.api.registration.IGuiHandlerRegistration;
 import mezz.jei.api.registration.IRecipeTransferRegistration;
 import mezz.jei.api.runtime.IJeiRuntime;
 import net.minecraft.resources.ResourceLocation;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 @JeiPlugin
 public class RefinedStorageModPlugin implements IModPlugin {
     private static final ResourceLocation ID = new ResourceLocation(Common.MOD_ID, "plugin");
-    private static final Logger LOGGER = LoggerFactory.getLogger(RefinedStorageModPlugin.class);
     @Nullable
     private static IJeiRuntime runtime;
 
@@ -36,54 +30,33 @@ public class RefinedStorageModPlugin implements IModPlugin {
 
     @Override
     public void onRuntimeAvailable(final IJeiRuntime newRuntime) {
-        if (runtime == null) {
-            initializePlatform(newRuntime.getJeiHelpers().getPlatformFluidHelper());
-        }
         RefinedStorageModPlugin.runtime = newRuntime;
     }
 
     @Override
+    public void onRuntimeUnavailable() {
+        RefinedStorageModPlugin.runtime = null;
+    }
+
+    @Override
     public void registerGuiHandlers(final IGuiHandlerRegistration registration) {
-        final RecipeModIngredientConverter converter = PlatformApi.INSTANCE.getIngredientConverter();
         registration.addGenericGuiContainerHandler(
             AbstractBaseScreen.class,
-            new ResourceGuiContainerHandler(converter, registration.getJeiHelpers().getIngredientManager())
+            new ResourceGuiContainerHandler(registration.getJeiHelpers().getIngredientManager())
         );
         registration.addGenericGuiContainerHandler(
             AbstractGridScreen.class,
-            new GridGuiContainerHandler(converter, registration.getJeiHelpers().getIngredientManager())
+            new GridGuiContainerHandler(registration.getJeiHelpers().getIngredientManager())
         );
         registration.addGenericGuiContainerHandler(
             AbstractBaseScreen.class,
             new ExclusionZonesGuiContainerHandler()
         );
-        registration.addGhostIngredientHandler(AbstractBaseScreen.class, new GhostIngredientHandler(converter));
+        registration.addGhostIngredientHandler(AbstractBaseScreen.class, new GhostIngredientHandler());
     }
 
     @Nullable
     public static IJeiRuntime getRuntime() {
         return runtime;
-    }
-
-    private void initializePlatform(final IPlatformFluidHelper<?> fluidHelper) {
-        LOGGER.debug("Enabling JEI integration");
-        registerIngredientConverters(fluidHelper);
-        registerGridSynchronizers();
-    }
-
-    private void registerGridSynchronizers() {
-        final JeiProxy jeiProxy = new JeiProxy();
-        PlatformApi.INSTANCE.getGridSynchronizerRegistry().register(
-            new ResourceLocation(Common.MOD_ID, "jei"),
-            new JeiGridSynchronizer(jeiProxy, false)
-        );
-        PlatformApi.INSTANCE.getGridSynchronizerRegistry().register(
-            new ResourceLocation(Common.MOD_ID, "jei_two_way"),
-            new JeiGridSynchronizer(jeiProxy, true)
-        );
-    }
-
-    private void registerIngredientConverters(final IPlatformFluidHelper<?> fluidHelper) {
-        PlatformApi.INSTANCE.addIngredientConverter(new JeiRecipeModIngredientConverter(fluidHelper));
     }
 }