+++ refinedstorage-jei-integration-fabric/src/main/java/com/refinedmods/refinedstorage/jei/fabric/JeiRefinedStoragePlugin.java
@@ -1,0 +1,13 @@
+package com.refinedmods.refinedstorage.jei.fabric;
+
+import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
+import com.refinedmods.refinedstorage2.platform.api.RefinedStoragePlugin;
+
+import static com.refinedmods.refinedstorage.jei.common.Common.init;
+
+public class JeiRefinedStoragePlugin implements RefinedStoragePlugin {
+    @Override
+    public void onPlatformApiAvailable(final PlatformApi platformApi) {
+        init(platformApi);
+    }
+}