+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/JeiHelper.java
@@ -2,17 +2,20 @@ package com.refinedmods.refinedstorage.jei.common;
 
 import mezz.jei.api.runtime.IJeiRuntime;
 
-public class JeiProxy {
-    public String getSearchFieldText() {
-        final IJeiRuntime runtime = RefinedStorageJeiModPlugin.getRuntime();
+public final class JeiHelper {
+    private JeiHelper() {
+    }
+
+    public static String getSearchFieldText() {
+        final IJeiRuntime runtime = RefinedStorageModPlugin.getRuntime();
         if (runtime == null) {
             return "";
         }
         return runtime.getIngredientFilter().getFilterText();
     }
 
-    public void setSearchFieldText(final String text) {
-        final IJeiRuntime runtime = RefinedStorageJeiModPlugin.getRuntime();
+    public static void setSearchFieldText(final String text) {
+        final IJeiRuntime runtime = RefinedStorageModPlugin.getRuntime();
         if (runtime != null) {
             runtime.getIngredientFilter().setFilterText(text);
         }