+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/JeiRecipeModIngredientConverter.java
@@ -7,15 +7,10 @@ import com.refinedmods.refinedstorage2.platform.api.support.resource.RecipeModIn
 import com.refinedmods.refinedstorage2.platform.common.support.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.common.support.resource.ItemResource;
 import mezz.jei.api.helpers.IPlatformFluidHelper;
+import mezz.jei.common.platform.Services;
 import net.minecraft.world.item.ItemStack;
 
 class JeiRecipeModIngredientConverter implements RecipeModIngredientConverter {
-    private final IPlatformFluidHelper<?> fluidHelper;
-
-    JeiRecipeModIngredientConverter(final IPlatformFluidHelper<?> fluidHelper) {
-        this.fluidHelper = fluidHelper;
-    }
-
     @Override
     public Optional<PlatformResourceKey> convertToResource(final Object ingredient) {
         final var fluid = Common.getPlatform().convertJeiIngredientToFluid(ingredient);
@@ -34,6 +29,7 @@ class JeiRecipeModIngredientConverter implements RecipeModIngredientConverter {
             return Optional.of(itemResource.toItemStack());
         }
         if (resource instanceof FluidResource fluidResource) {
+            final IPlatformFluidHelper<?> fluidHelper = Services.PLATFORM.getFluidHelper();
             return Optional.of(fluidHelper.create(
                 fluidResource.fluid(),
                 fluidHelper.bucketVolume(),